requirements.txt
Flask==3.0.2
pycryptodome==3.20.0